fix: types for cover letter
