+++ src/templates/coveringLetter/sample.ts
@@ -1,19 +1,4 @@
-import { Document } from "@govtechsg/decentralized-renderer-react-components";
-
-export interface CoveringLetter extends Document {
-  name: string;
-  logo?: string;
-  title?: string;
-  remarks?: string;
-  titleColor?: string;
-  backgroundColor?: string;
-  remarksColor?: string;
-  links?: {
-    self: {
-      href: string;
-    };
-  };
-}
+import { CoveringLetter } from "./types";
 
 export const coveringLetter: CoveringLetter = {
   name: "Covering Letter",
